refactor(onboard): extract onboard-ollama-proxy.ts
Move the entire Ollama auth-proxy subsystem into a dedicated module:
token/PID persistence, proxy spawn/kill lifecycle, model prompt/pull,
and the exposure warning.

Zero coupling to other onboard logic — all dependencies (runner, ports,
local-inference, credentials, model-prompts) are imported directly.
