+++ src/lib/onboard-ollama-proxy.ts
@@ -1,0 +1,270 @@
+// @ts-nocheck
+// SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+//
+// Ollama auth-proxy lifecycle: token persistence, PID management,
+// proxy start/stop, model pull and validation.
+
+const fs = require("fs");
+const os = require("os");
+const path = require("path");
+const { spawn, spawnSync } = require("child_process");
+const { ROOT, SCRIPTS, run, runCapture, shellQuote } = require("./runner");
+const { OLLAMA_PORT, OLLAMA_PROXY_PORT } = require("./ports");
+const {
+  getDefaultOllamaModel,
+  getBootstrapOllamaModelOptions,
+  getOllamaModelOptions,
+  getOllamaWarmupCommand,
+  validateOllamaModel,
+} = require("./local-inference");
+const { prompt } = require("./credentials");
+const { promptManualModelId } = require("./model-prompts");
+
+// ── State ────────────────────────────────────────────────────────
+
+const PROXY_STATE_DIR = path.join(os.homedir(), ".nemoclaw");
+const PROXY_TOKEN_PATH = path.join(PROXY_STATE_DIR, "ollama-proxy-token");
+const PROXY_PID_PATH = path.join(PROXY_STATE_DIR, "ollama-auth-proxy.pid");
+
+let ollamaProxyToken: string | null = null;
+
+function sleep(seconds) {
+  spawnSync("sleep", [String(seconds)]);
+}
+
+// ── Proxy state dir ──────────────────────────────────────────────
+
+function ensureProxyStateDir(): void {
+  if (!fs.existsSync(PROXY_STATE_DIR)) {
+    fs.mkdirSync(PROXY_STATE_DIR, { recursive: true });
+  }
+}
+
+// ── Token persistence ────────────────────────────────────────────
+
+function persistProxyToken(token: string): void {
+  ensureProxyStateDir();
+  fs.writeFileSync(PROXY_TOKEN_PATH, token, { mode: 0o600 });
+  // mode only applies on creation; ensure permissions on existing files too
+  fs.chmodSync(PROXY_TOKEN_PATH, 0o600);
+}
+
+function loadPersistedProxyToken(): string | null {
+  try {
+    if (fs.existsSync(PROXY_TOKEN_PATH)) {
+      const token = fs.readFileSync(PROXY_TOKEN_PATH, "utf-8").trim();
+      return token || null;
+    }
+  } catch {
+    /* ignore */
+  }
+  return null;
+}
+
+// ── PID persistence ──────────────────────────────────────────────
+
+function persistProxyPid(pid: number | null | undefined): void {
+  if (!Number.isInteger(pid) || pid <= 0) return;
+  ensureProxyStateDir();
+  fs.writeFileSync(PROXY_PID_PATH, `${pid}\n`, { mode: 0o600 });
+  fs.chmodSync(PROXY_PID_PATH, 0o600);
+}
+
+function loadPersistedProxyPid(): number | null {
+  try {
+    if (!fs.existsSync(PROXY_PID_PATH)) return null;
+    const raw = fs.readFileSync(PROXY_PID_PATH, "utf-8").trim();
+    const pid = Number.parseInt(raw, 10);
+    return Number.isInteger(pid) && pid > 0 ? pid : null;
+  } catch {
+    return null;
+  }
+}
+
+function clearPersistedProxyPid(): void {
+  try {
+    if (fs.existsSync(PROXY_PID_PATH)) {
+      fs.unlinkSync(PROXY_PID_PATH);
+    }
+  } catch {
+    /* ignore */
+  }
+}
+
+// ── Process management ───────────────────────────────────────────
+
+function isOllamaProxyProcess(pid: number | null | undefined): boolean {
+  if (!Number.isInteger(pid) || pid <= 0) return false;
+  const cmdline = runCapture(["ps", "-p", String(pid), "-o", "args="], { ignoreError: true });
+  return Boolean(cmdline && cmdline.includes("ollama-auth-proxy.js"));
+}
+
+function spawnOllamaAuthProxy(token: string): number | null {
+  const child = spawn(process.execPath, [path.join(SCRIPTS, "ollama-auth-proxy.js")], {
+    detached: true,
+    stdio: "ignore",
+    env: {
+      ...process.env,
+      OLLAMA_PROXY_TOKEN: token,
+      OLLAMA_PROXY_PORT: String(OLLAMA_PROXY_PORT),
+      OLLAMA_BACKEND_PORT: String(OLLAMA_PORT),
+    },
+  });
+  child.unref();
+  persistProxyPid(child.pid);
+  return child.pid ?? null;
+}
+
+function killStaleProxy(): void {
+  try {
+    const persistedPid = loadPersistedProxyPid();
+    if (isOllamaProxyProcess(persistedPid)) {
+      run(["kill", String(persistedPid)], { ignoreError: true, suppressOutput: true });
+    }
+    clearPersistedProxyPid();
+
+    // Best-effort cleanup for older proxy processes created before the PID file
+    // existed. Only kill processes that are actually the auth proxy, not
+    // unrelated services that happen to use the same port.
+    const pidOutput = runCapture(["lsof", "-ti", `:${OLLAMA_PROXY_PORT}`], { ignoreError: true });
+    if (pidOutput && pidOutput.trim()) {
+      for (const pid of pidOutput.trim().split(/\s+/)) {
+        if (isOllamaProxyProcess(Number.parseInt(pid, 10))) {
+          run(["kill", pid], { ignoreError: true, suppressOutput: true });
+        }
+      }
+      sleep(1);
+    }
+  } catch {
+    /* ignore */
+  }
+}
+
+// ── Public API ───────────────────────────────────────────────────
+
+function startOllamaAuthProxy(): void {
+  const crypto = require("crypto");
+  killStaleProxy();
+
+  ollamaProxyToken = crypto.randomBytes(24).toString("hex");
+  // Don't persist yet — wait until provider is confirmed in setupInference.
+  // If the user backs out to a different provider, the token stays in memory
+  // only and is discarded.
+  const pid = spawnOllamaAuthProxy(ollamaProxyToken);
+  sleep(1);
+  if (!isOllamaProxyProcess(pid)) {
+    console.error(`  Warning: Ollama auth proxy did not start on :${OLLAMA_PROXY_PORT}`);
+  }
+}
+
+/**
+ * Ensure the auth proxy is running — called on sandbox connect to recover
+ * from host reboots where the background proxy process was lost.
+ */
+function ensureOllamaAuthProxy(): void {
+  // Try to load persisted token first — if none, this isn't an Ollama setup.
+  const token = loadPersistedProxyToken();
+  if (!token) return;
+
+  const pid = loadPersistedProxyPid();
+  if (isOllamaProxyProcess(pid)) {
+    ollamaProxyToken = token;
+    return;
+  }
+
+  // Proxy not running — restart it with the persisted token.
+  killStaleProxy();
+  ollamaProxyToken = token;
+  spawnOllamaAuthProxy(token);
+  sleep(1);
+}
+
+function getOllamaProxyToken(): string | null {
+  if (ollamaProxyToken) return ollamaProxyToken;
+  // Fall back to persisted token (resume / reconnect scenario)
+  ollamaProxyToken = loadPersistedProxyToken();
+  return ollamaProxyToken;
+}
+
+async function promptOllamaModel(gpu = null) {
+  const installed = getOllamaModelOptions();
+  const options = installed.length > 0 ? installed : getBootstrapOllamaModelOptions(gpu);
+  const defaultModel = getDefaultOllamaModel(gpu);
+  const defaultIndex = Math.max(0, options.indexOf(defaultModel));
+
+  console.log("");
+  console.log(installed.length > 0 ? "  Ollama models:" : "  Ollama starter models:");
+  options.forEach((option, index) => {
+    console.log(`    ${index + 1}) ${option}`);
+  });
+  console.log(`    ${options.length + 1}) Other...`);
+  if (installed.length === 0) {
+    console.log("");
+    console.log("  No local Ollama models are installed yet. Choose one to pull and load now.");
+  }
+  console.log("");
+
+  const choice = await prompt(`  Choose model [${defaultIndex + 1}]: `);
+  const index = parseInt(choice || String(defaultIndex + 1), 10) - 1;
+  if (index >= 0 && index < options.length) {
+    return options[index];
+  }
+  return promptManualModelId("  Ollama model id: ", "Ollama");
+}
+
+function printOllamaExposureWarning() {
+  console.log("");
+  console.log("  ⚠ Ollama is binding to 0.0.0.0 so the sandbox can reach it via Docker.");
+  console.log("    This exposes the Ollama API to your local network (no auth required).");
+  console.log("    On public WiFi, any device on the same network can send prompts to your GPU.");
+  console.log("    See: CNVD-2025-04094, CVE-2024-37032");
+  console.log("");
+}
+
+function pullOllamaModel(model) {
+  const result = spawnSync("bash", ["-c", `ollama pull ${shellQuote(model)}`], {
+    cwd: ROOT,
+    encoding: "utf8",
+    stdio: "inherit",
+    timeout: 600_000,
+    env: { ...process.env },
+  });
+  if (result.signal === "SIGTERM") {
+    console.error(
+      `  Model pull timed out after 10 minutes. Try a smaller model or check your network connection.`,
+    );
+    return false;
+  }
+  return result.status === 0;
+}
+
+function prepareOllamaModel(model, installedModels = []) {
+  const alreadyInstalled = installedModels.includes(model);
+  if (!alreadyInstalled) {
+    console.log(`  Pulling Ollama model: ${model}`);
+    if (!pullOllamaModel(model)) {
+      return {
+        ok: false,
+        message:
+          `Failed to pull Ollama model '${model}'. ` +
+          "Check the model name and that Ollama can access the registry, then try another model.",
+      };
+    }
+  }
+
+  console.log(`  Loading Ollama model: ${model}`);
+  run(getOllamaWarmupCommand(model), { ignoreError: true });
+  return validateOllamaModel(model);
+}
+
+module.exports = {
+  ensureOllamaAuthProxy,
+  getOllamaProxyToken,
+  persistProxyToken,
+  startOllamaAuthProxy,
+  promptOllamaModel,
+  printOllamaExposureWarning,
+  pullOllamaModel,
+  prepareOllamaModel,
+};
